refactor togglProject
Tasks are deleted and entrysums are directly in the project.
